Register/requirements.txt
Telethon
PySocks
PyMySQL
requests
flask